Encapsulate coordinates of garages into ParkingLocation object for feature scalability
Created object to relate location, restricted dates/times, and decal requirements and ensured that it worked as it did before this change.

Bug: when selecting closest location, should make the marker for that location green (or just always make the 'selected location' green)

Will need to add the real data for these parking garages and filter based of them, after that we can figure out a better way to store them in the code and add the rest of the smaller lots.